% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj_amt_indiffs}
\alias{adj_amt_indiffs}
\title{Indifference points from adjusting amount procedure}
\usage{
adj_amt_indiffs(data, block_indic = "del", order_indic = NULL)
}
\arguments{
\item{data}{A dataframe where each row corresponds to a binary choice, with at least columns \code{val_imm}, \code{val_del}, and \code{imm_chosen}, along with a block indicator and (if applicable) an order indicator.}

\item{block_indic}{Column name of the block indicator---i.e., the column that will identify a block of trials for which an indifference point should be computed. If unspecified, defaults to \code{'del'}, which assumes that each block corresponds to a different delay.}

\item{order_indic}{Column name of the order indicator---i.e., the column that specifies the order in which trials were completed. Sorting by this column within a block should sort the rows in chronological order. If unspecified, the rows are assumed to already be in chronological order.}
}
\value{
A dataframe with two columns: one for the block indicator and another for the corresponding indifference point
}
\description{
Compute indifference points for data from an adjusting amount procedure (also called a "titrating procedure").
}
\examples{
\dontrun{
data("adj_amt_sim")
adj_amt_indiffs(adj_amt_sim)
adj_amt_indiffs(adj_amt_sim, block_indic = 'del', order_indic = 'trial_idx')
}
}
