% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{adj_amt_sim}
\alias{adj_amt_sim}
\title{Simulated adjusting amount procedure}
\description{
A minimal example of data from a single participant for an adjusting amount procedure.
}
\author{
Isaac Kinley \email{isaac.kinley@gmail.com}
}
\keyword{adjusting}
\keyword{amount}
\keyword{choice}
\keyword{delay}
\keyword{discounting}
\keyword{simulated}
\keyword{temporal}
\keyword{titrating}
