% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.td_bcm}
\alias{coef.td_bcm}
\title{Extract model coefficients}
\usage{
\method{coef}{td_bcm}(mod)
}
\value{
A named vector of coefficients
}
\description{
Get coefficients of a temporal discounting binary choice model
}
