% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.td_bcnm}
\alias{coef.td_bcnm}
\title{Extract model coefficients}
\usage{
\method{coef}{td_bcnm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{td_bcnm}.}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of coefficients.
}
\description{
Get coefficients of a temporal discounting binary choice nonlinear model.
}
\seealso{
Other nonlinear binary choice model functions: 
\code{\link{deviance.td_bcnm}()},
\code{\link{fitted.td_bcnm}()},
\code{\link{logLik.td_bcnm}()},
\code{\link{predict.td_bcnm}()},
\code{\link{residuals.td_bcnm}()},
\code{\link{td_bcnm}()}
}
\concept{nonlinear binary choice model functions}
