% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.td_gnlm}
\alias{coef.td_gnlm}
\title{Extract coefficients}
\usage{
\method{coef}{td_gnlm}(mod, bounded = T)
}
\arguments{
\item{bounded}{Boolean specifying whether parameters should be bounded (e.g., k > 0) or should be unbounded. Boended and unbounded coefficients are related by some transform. Bounded coefficients can be plugged directly into discount functions; optimization is performed on unbounded coefficients.}
}
\value{
A names vector of coefficients
}
\description{
Get coefficients of delay discounting function
}
