% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-model-fitting.R
\name{dd_det_model}
\alias{dd_det_model}
\title{Deterministic delay discounting model}
\usage{
dd_det_model(data, discount_function = "all", param_ranges = NULL, silent = T)
}
\arguments{
\item{data}{A data frame with columns `indiff` for the pre-computed indifference points and `del` for the delay}

\item{discount_function}{A vector of strings specifying the name of the discount functions to use. Options are `'hyperbolic'`, `'exponential'`, `'inverse-q-exponential'`, `'nonlinear-time-hyperbolic'`, `'scaled-exponential'`, `'dual-systems-exponential'`, and `'nonlinear-time-exponential'`. Default is `'all'`, meaning every discount function is tested and the one with the best AIC is selected.}

\item{param_ranges}{A list containing the starting values to try for each parameter. Defaults to `c(-5, 0, 5)` for most parameters}

\item{silent}{A Boolean specifying whether the call to `optim` (which occurs in a `try` block) should be silent on error}
}
\value{
A list from `optim` with additional components specifying the AIC, the ED50, the discount function, and the probabilistic model
}
\description{
Compute a deterministic model for a single subject's delay discounting
}
\note{
The `par` component of the output list is for internal use. For statistical analyses, use the `untransformed_parameters`. `par` contains the parameters after various transformations intended to keep them within certain bounds (e.g., k parameters should never be negative)
}
\examples{
# Generate data
df <- data.frame(del = exp(1:10), indiff = 1 / (1 + 0.001*exp(1:10)))
# Fit model
mod <- dd_det_model(df)
print(mod$discount_function_name)
}
