% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{deviance.td_bcnm}
\alias{deviance.td_bcnm}
\title{Model deviance}
\usage{
\method{deviance}{td_bcnm}(mod)
}
\arguments{
\item{mod}{An object of class \code{td_bcnm}}
}
\description{
Compute deviance for a temporal discounting binary choice model.
}
