% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{deviance.td_ddm}
\alias{deviance.td_ddm}
\title{Model deviance}
\usage{
\method{deviance}{td_ddm}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{td_ddm}}

\item{...}{Additional arguments currently not used.}
}
\description{
Compute deviance for a temporal discounting drift diffusion model
}
\seealso{
Other drift diffusion model functions: 
\code{\link{coef.td_ddm}()},
\code{\link{fitted.td_ddm}()},
\code{\link{logLik.td_ddm}()},
\code{\link{predict.td_ddm}()},
\code{\link{td_ddm}()}
}
\concept{drift diffusion model functions}
