% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{discount_function}
\alias{discount_function}
\title{Get discount function from model}
\usage{
discount_function(mod)
}
\arguments{
\item{mod}{A temporal discounting model.}
}
\value{
The name of the discount function.
}
\description{
Access the name of the discount function of a model.
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bcnm(td_bc_single_ptpt)
discount_function(mod)
}
}
