% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ED50}
\alias{ED50}
\title{Median effective delay}
\usage{
ED50(mod, val_del = NULL)
}
\arguments{
\item{mod}{A temporal discounting model.}

\item{val_del}{Delayed value, if applicable (i.e., if magnitude effects are accounted for).}
}
\value{
Median effective delay value.
}
\description{
Compute the median effective delay.
}
\examples{
\donttest{
data("td_bc_single_ptpt")
mod <- td_bcnm(td_bc_single_ptpt, discount_function = "exponential")
print(ED50(mod))
}
}
