% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.td_bcm}
\alias{fitted.td_bcm}
\title{Get fitted values}
\usage{
\method{fitted}{td_bcm}(mod)

\method{fitted}{td_bcm}(mod)
}
\value{
A named vector of fitted values

A named vector of fitted values
}
\description{
Get fitted values of a temporal discounting binary choice model

Get fitted values of a temporal discounting indifference point model
}
