% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_discount_functions}
\alias{get_available_discount_functions}
\title{Get all available pre-defined discount functions}
\usage{
get_available_discount_functions()
}
\value{
A character vector containing the names of the avialable pre-defined discount functions.
}
\description{
Get a list of all of the available pre-defined discount functions.
}
\examples{
\donttest{
get_available_discount_functions()
}
}
