% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{invariance_checks}
\alias{invariance_checks}
\title{Check for invariant responding}
\usage{
invariance_checks(data, warn = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} with columns \code{val_imm}, \code{val_del} and \code{del_chosen}, representing data from a single participant.}

\item{warn}{Logical: give a warning for invariant responding?}
}
\value{
Named vector specifying whether the participant chose only immediate rewards (\code{all_imm}) or chose all delayed rewards (\code{all_del}).
}
\description{
Check whether participants always chose the immediate reward or always chose the delayed reward.
}
\examples{
\donttest{
# On a model
data("td_bc_single_ptpt")
attention_checks(td_bc_single_ptpt)
}
}
