% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kirby_consistency}
\alias{kirby_consistency}
\title{Compute consistency score}
\usage{
kirby_consistency(
  data,
  discount_function = c("hyperbolic", "exponential", "power")
)
}
\arguments{
\item{data}{Responses to score.}

\item{discount_function}{Should \eqn{k} values be computed according to the hyperbolic, exponential, or power discount function? The original method uses the hyperbolic, but in principle the exponential and power are also possible (though these should be considered experimental features).}
}
\value{
A consistency score between 0 and 1.
}
\description{
Compute the consistency score per the method of Kirby et al. (1999, \doi{10.1037//0096-3445.128.1.78}). This is described in detail in Kaplan et al. (2016, \doi{10.1007/s40614-016-0070-9}), where it's suggested that a consistency score below 0.75 might be a sign of inattentive responding.
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- kirby_consistency(td_bc_single_ptpt)
}
}
