% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.td_bcm}
\alias{logLik.td_bcm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{td_bcm}(mod)
}
\arguments{
\item{mod}{An object of class \code{td_bcm}}
}
\description{
Compute log-likelihood for a temporal discounting binary choice model.
}
