% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.td_ddm}
\alias{logLik.td_ddm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{td_ddm}(mod, type = c("resp_rt", "resp", "rt"))
}
\arguments{
\item{mod}{An object of class \code{td_bcnm}}

\item{type}{Should probabilities /probability densities be computed for responses and RTs (\code{'resp_rt'}, default) or responses only (\code{'resp'})?}
}
\description{
Compute log-likelihood for a temporal discounting drift diffusion model
}
