% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.td_ipm}
\alias{logLik.td_ipm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{td_ipm}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{td_ipm}}

\item{...}{Additional arguments currently not used.}
}
\description{
Compute log-likelihood for a temporal discounting indifference point model.
}
\seealso{
Other indifference point model functions: 
\code{\link{coef.td_ipm}()},
\code{\link{fitted.td_ipm}()},
\code{\link{predict.td_ipm}()},
\code{\link{residuals.td_ipm}()}
}
\concept{indifference point model functions}
