% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{most_consistent_indiffs}
\alias{most_consistent_indiffs}
\title{Experimental method for computing indifference points}
\usage{
most_consistent_indiffs(data)
}
\arguments{
\item{data}{Responses to score}
}
\description{
Experimental method for computing indifference points
}
