% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.td_um}
\alias{plot.td_um}
\title{Plot models}
\usage{
\method{plot}{td_um}(x, ...)
}
\arguments{
\item{x}{A delay discounting model. See \code{dd_prob_model} and \code{dd_det_model}}

\item{...}{Additional arguments to \code{plot()}}

\item{type}{Type of plot to generate}
}
\description{
Plot delay discounting models
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
plot(mod, type = 'summary')
plot(mod, type = 'endpoints')
}
}
