% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.tdbcm}
\alias{plot.tdbcm}
\title{Plot models}
\usage{
\method{plot}{tdbcm}(mod, type = c("summary", "endpoints", "scores"), ...)
}
\arguments{
\item{mod}{A delay discounting model. See `dd_prob_model` and `dd_det_model`}

\item{type}{Type of plot to generate}

\item{p_range}{In addition to the indifference points (where the probability of choosing the immediate reward is 0.5), other "preference points" can also be plotted (e.g., points where the probability of choosing the immediate reward is 0.8). This allows visualization of how stochastic decision making is. Defaults to `c(0.4, 0.6)`}
}
\description{
Plot delay discounting models
}
\examples{
# Generate data
df <- data.frame(val_imm = seq(1, 99, length.out = 10), val_del = 100, del = rep(exp(1:10), each=10))
logistic <- function(x) 1 / (1 + exp(-x))
logit <- function(x) log(x / (1 - x))
prob <- logistic(logit(df$val_imm / df$val_del) - logit(1 / (1 + 0.001*df$del)))
df$imm_chosen <- runif(nrow(df)) < prob
# Fit model
mod <- dd_prob_model(df)
plot_dd(mod)
}
