% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_choices}
\alias{plot_choices}
\title{Plot choices}
\usage{
plot_choices(data, legend = TRUE, ...)
}
\arguments{
\item{data}{A data frame with columns \code{val_imm} and \code{val_del} for the values of the immediate and delayed rewards, \code{del} for the delay, and \code{imm_chosen} (Boolean) for whether the immediate reward was chosen.}

\item{legend}{Logical: display a legend?}

\item{...}{Additional arguments to \code{plot()}.}
}
\value{
No return value (called to produce a plot)
}
\description{
Create a plot that displays binary choices at each delay
}
\examples{
\donttest{
data('td_bc_single_ptpt')
plot_choices(td_bc_single_ptpt)
}
}
