% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.td_ipm}
\alias{predict.td_ipm}
\title{Model Predictions}
\usage{
\method{predict}{td_ipm}(object, ...)
}
\arguments{
\item{object}{A temporal discounting indifference point model. See \code{td_ipm}.}

\item{...}{Additional arguments currently not used.}

\item{del}{Vector of delays for which to predict indifference points. If omitted, the data used to fit the model will be used for prediction.}

\item{newdata}{Optionally, a data frame to use for prediction. This overrides the \code{del} argument.}
}
\value{
A vector of predictions
}
\description{
Generate predictions from a temporal discounting indifference point model
}
\examples{
\dontrun{
data("td_ip_simulated_ptpt")
mod <- td_ipm(td_ip_simulated_ptpt, discount_function = 'hyperbolic')
indiffs <- predict(mod, newdata = data.frame(del = 1:100), type = 'indiff')
}
}
