% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.tdbclm}
\alias{predict.tdbclm}
\title{Model Predictions}
\usage{
\method{predict}{tdbclm}(mod, newdata = NULL, type = "indiff", ...)
}
\arguments{
\item{mod}{A temporal discounting binary choice linear model. See `td_bclm`.}

\item{newdata}{Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.}

\item{type}{The type of prediction required. For `"indiff"` (default) gives predicted indifference points. In this case, `newdata` needs only a `del` column. For all other values (e.g. `"link"`, `"response"`), this function is just a wrapper to `predict.glm()`}

\item{...}{Additional arguments passed to predict.glm}
}
\value{
A vector of predictions
}
\description{
Generate predictions from a temporal discounting binary choice linear model
}
