% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.tdbcm}
\alias{predict.tdbcm}
\title{Model Predictions}
\usage{
\method{predict}{tdbcm}(mod, newdata = NULL, type = c("link", "response", "indiff"))
}
\arguments{
\item{mod}{A temporal discounting model. See `tdbcm`.}

\item{newdata}{Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.}

\item{type}{The type of prediction required. As in predict.glm, `"link"` (default) and `"response"` give predictions on the scales of the linear predictors and response variable, respectively. `"indiff"` gives predicted indifference points. In this case, `newdata` needs only a `del` column.}
}
\value{
A vector of predictions
}
\description{
Generate predictions from a temporal discounting binary choice model
}
