% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-model-fitting.R
\name{predict_indiffs}
\alias{predict_indiffs}
\title{Predict indifference points}
\usage{
predict_indiffs(mod, del = NULL)
}
\arguments{
\item{mod}{A probabilistic delay discounting model. See `dd_prob_model`}

\item{del}{A vector of delays. Defaults to the delays from the data used to fit the model}
}
\value{
A vector of indifference points
}
\description{
Predict the indifference point at a given delay
}
\examples{
\dontrun{
df <- data.frame(val_imm = seq(1, 99, length.out = 10), val_del = 100, del = rep(exp(1:10), each=10))
logistic <- function(x) 1 / (1 + exp(-x))
logit <- function(x) log(x / (1 - x))
prob <- logistic(logit(df$val_imm / df$val_del) - logit(1 / (1 + 0.001*df$del)))
df$imm_chosen <- runif(nrow(df)) < prob
mod <- dd_prob_model(df)
indiffs <- predict_indiffs(mod)
plot(indiffs ~ log(mod$data$del), type = 'l')
}
}
