% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear-model-fitting.R
\name{predict_prob_imm}
\alias{predict_prob_imm}
\title{Predict decision probabilities}
\usage{
predict_prob_imm(mod, data = NULL)
}
\arguments{
\item{mod}{A probabilistic delay discounting model. See `dd_prob_model`}

\item{data}{A data frame with columns `val_imm`, `val_del`, and `del`, specifying the immediate and delayed rewards, and the delay of the delayed reward. Defaults to the data used to fit the model}
}
\value{
A vector of probabilities
}
\description{
Predict the probability of selecting the immediate reward
}
\examples{
# Generate data
df <- data.frame(val_imm = seq(1, 99, length.out = 10), val_del = 100, del = rep(exp(1:10), each=10))
logistic <- function(x) 1 / (1 + exp(-x))
logit <- function(x) log(x / (1 - x))
prob <- logistic(logit(df$val_imm / df$val_del) - logit(1 / (1 + 0.001*df$del)))
df$imm_chosen <- runif(nrow(df)) < prob
# Fit model
mod <- dd_prob_model(df)
prob_imm <- predict_prob_imm(mod)
boxplot(prob_imm ~ mod$data$imm_chosen)
}
