% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.td_bcm}
\alias{print.td_bcm}
\title{#' @export
AIC.tdm <- function(mod, k = 2) {
  return(-2*as.numeric(logLik(mod)) + k*length(coef(mod)))
}}
\usage{
\method{print}{td_bcm}(mod)
}
\description{
#' @export
BIC.tdm <- function(mod) {
  return(AIC(mod, k = log(nrow(mod$data))))
}
}
