% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.td_bcnm}
\alias{residuals.td_bcnm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{td_bcnm}(object, type = c("deviance", "pearson", "response"), ...)
}
\arguments{
\item{object}{A temporal discounting binary choice model. See \code{td_bcnm}.}

\item{type}{The type of residuals to be returned. See \code{residuals.glm}.}

\item{...}{Additional arguments currently not used.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting binary choice model
}
\seealso{
Other nonlinear binary choice model functions: 
\code{\link{coef.td_bcnm}()},
\code{\link{deviance.td_bcnm}()},
\code{\link{fitted.td_bcnm}()},
\code{\link{logLik.td_bcnm}()},
\code{\link{predict.td_bcnm}()},
\code{\link{td_bcnm}()}
}
\concept{nonlinear binary choice model functions}
