% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.td_gnlm}
\alias{residuals.td_gnlm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{td_gnlm}(mod, type = c("deviance", "pearson", "response"))
}
\arguments{
\item{mod}{A temporal discounting model. See `td_gnlm`.}

\item{type}{The type of residuals to be returned. See `residuals.glm`.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting model
}
