% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.td_ipm}
\alias{residuals.td_ipm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{td_ipm}(mod, type = c("response", "pearson"))
}
\arguments{
\item{mod}{A temporal discounting model. See `td_bcm`.}

\item{type}{The type of residuals to be returned. See `residuals.nls`.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting indifference point model
}
