% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.tdbcm}
\alias{residuals.tdbcm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{tdbcm}(mod, type = c("deviance", "pearson", "response"))
}
\arguments{
\item{mod}{A temporal discounting model. See `tdbcm`.}

\item{type}{The type of residuals to be returned. See `residuals.glm`.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting model
}
