% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.tdipm}
\alias{residuals.tdipm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{tdipm}(mod, type = "response")
}
\arguments{
\item{mod}{A temporal discounting model. See `tdbcm`.}

\item{type}{The type of residuals to be returned. See `residuals.glm`.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting indifference point model
}
