% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{td_bc_single_ptpt}
\alias{td_bc_single_ptpt}
\title{Binary choice data for a single participant}
\description{
70 binary choices made by a single participant. Along with the columns required by \code{td_bcnm}, the reaction time (\code{rt}) is recorded.
}
\references{
\url{https://doi.org/10.31234/osf.io/y2fdh}
}
\author{
Isaac Kinley \email{isaac.kinley@gmail.com}
}
\keyword{choice}
\keyword{delay}
\keyword{discounting}
\keyword{temporal}
