% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_bclm.R
\name{td_bclm}
\alias{td_bclm}
\title{Temporal discounting binary choice linear model}
\usage{
td_bclm(
  data,
  model = c("hyperbolic.1", "hyperbolic.2", "exponential.1", "exponential.2",
    "scaled-exponential.1", "nonlinear-time-hyperbolic.2",
    "nonlinear-time-exponential.2"),
  ...
)
}
\arguments{
\item{model}{A string specifying which model to use. Below is a list of these models' linear predictors and the means by which we can recover discount function parameters. \cr
\code{'hyperbolic.1'}: \eqn{\beta_1(1 - v_D/v_I) + \beta_2 t}; \eqn{k = \beta_2/\beta_1} \cr
\code{'hyperbolic.2'}: \eqn{\beta_1(\sigma^{-1}[v_I/v_D] + \log t) + \beta_2}; \eqn{k = \exp[\beta_2/\beta_1]} \cr
\code{'exponential.1'}: \eqn{\beta_1 \log (v_I/v_D) + \beta_2 t}; \eqn{k = \beta_2/\beta_1} \cr
\code{'exponential.2'}: \eqn{\beta_1(G^{-1}[v_I/v_D] + \log t) + \beta_2}; \eqn{k = \exp[\beta_2/\beta_1]} \cr
\code{'scaled-exponential.1'}: \eqn{\beta_1 \log (v_I/v_D) + \beta_2 t + \beta_3}; \eqn{k = \beta_2/\beta_1}, \eqn{w = \exp[-\beta_3/\beta_1]}
\code{'nonlinear-time-hyperbolic.2'}: \eqn{\beta_1(\sigma^{-1}[v_I/v_D]) + \beta_2\log t + \beta_3}; \eqn{k = \exp[\beta_3/\beta_1]}, \eqn{s = \beta_2/\beta_1} \cr
\code{'nonlinear-time-hyperbolic.2'}: \eqn{\beta_1(G^{-1}[v_I/v_D]) + \beta_2\log t + \beta_3}; \eqn{k = \exp[\beta_3/\beta_1]}, \eqn{s = \beta_2/\beta_1} \cr
where \eqn{\sigma^{-1}[\cdot]} is the quantile function of the standard logistic distribution \eqn{G^{-1}[\cdot]} is the quantile function of the standard Gumbel distribution}

\item{...}{Additional arguments passed to \code{glm}}
}
\value{
An object of class \code{td_bclm}, nearly identical to a \code{glm} but with an additional \code{config} component.
}
\description{
Compute a binary choice linear model for a single subject. In these models, we can recover the parameters of a discount function from the weights of a standard logistic regression.
\eqn{\beta_1}
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
print(coef(mod))
}
}
