% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_bcm.R
\name{td_bcm}
\alias{td_bcm}
\title{Temporal discounting binary choice model}
\usage{
td_bcm(
  data,
  discount_function = c("all", "hyperbolic", "exponential", "inverse-q-exponential",
    "nonlinear-time-hyperbolic", "scaled-exponential", "dual-systems-exponential",
    "nonlinear-time-exponential", "model-free", "noise"),
  choice_rule = c("logistic", "probit", "power"),
  fixed_ends = F,
  fit_err_rate = F,
  gamma_par_starts = c(0.01, 1, 100),
  eps_par_starts = c(0.01, 0.25),
  silent = T,
  optim_args = list(),
  ...
)
}
\arguments{
\item{data}{A data frame with columns `val_imm` and `val_del` for the values of the immediate and delayed rewards, `del` for the delay, and `imm_chosen` (Boolean) for whether the immediate reward was chosen. Other columns can also be present but will be ignored.}

\item{discount_function}{A string specifying the name of the discount functions to use, or an object of class `td_fn` (used for creating custom discount functions). See `td_fn()`.}

\item{choice_rule}{A string specifying whether the `'logistic'` (default), `'probit'`, or `'power'` choice rule should be used.}

\item{fixed_ends}{A Boolean (false by default) specifying whether the model should satisfy the desiderata that subjects should always prefer something over nothing (i.e., nonzero delayed reward over nothing) and the same reward sooner rather than later.}

\item{fit_err_rate}{A Boolean (false by default) specifying whether the model should include an error rate (parameterized by "eps"). See Eq. 5 here: https://doi.org/10.3758/s13428-015-0672-2}

\item{gamma_par_starts}{A vector of starting values to try for the "gamma" parameter (which controls the steepness of the choice rule) during optimization.}

\item{eps_par_starts}{A vector of starting values to try for the "eps" parameter (which controls the error rate) during optimization. Ignored if `fit_err_rate = FALSE`.}

\item{silent}{Boolean (true by default). The call to `optim()` occurs within a `try()` wrapper. The value of `silent` is passed along to `try()`.}

\item{optim_args}{Additional arguments to pass to `optim()`. Default is `list(silent = T)`.}
}
\value{
An object of class `td_bcm` with components `data` (containing the data used for fitting), `config` (containing the internal configuration of the model, including the `discount_function`), and `optim` (the output of `optim()`).
}
\description{
Compute a binary choice model for a single subject
}
