% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{td_ip_simulated_ptpt}
\alias{td_ip_simulated_ptpt}
\title{Simulated indifference point data for a single participant}
\description{
A dataframe containing simulated indifference points from a single participant exhibiting approximately hyperbolic discounting.
}
\author{
Isaac Kinley \email{isaac.kinley@gmail.com}
}
\keyword{choice}
\keyword{delay}
\keyword{discounting}
\keyword{simulated}
\keyword{temporal}
