% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_ipm.R
\name{td_ipm}
\alias{td_ipm}
\title{Indifference point model}
\usage{
td_ipm(
  data,
  discount_function = c("all", "hyperbolic", "exponential", "inverse-q-exponential",
    "nonlinear-time-hyperbolic", "scaled-exponential", "dual-systems-exponential",
    "nonlinear-time-exponential", "model-free", "noise"),
  optim_args = list(),
  silent = T
)
}
\arguments{
\item{data}{A data frame with columns `indiff` for the pre-computed indifference points and `del` for the delay}

\item{discount_function}{A vector of strings specifying the name of the discount functions to use, or an object of class `td_fn`.}

\item{optim_args}{A list of additional args to pass to `optim`}

\item{silent}{A Boolean specifying whether the call to `optim` (which occurs in a `try` block) should be silent on error}
}
\value{
A list from `optim` with additional components specifying the AIC, the ED50, the discount function, and the probabilistic model
}
\description{
Compute a model of a single subject's indifference points
}
