% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AUC}
\alias{AUC}
\title{Area under the curve (AUC)}
\usage{
AUC(mod, min_del = 0, max_del = NULL, val_del = NULL, verbose = T, ...)
}
\arguments{
\item{mod}{A temporal discounting model. See `td_gnlm`}

\item{min_del}{Lower limit to use for integration}

\item{max_del}{Upper limit to use for integration}

\item{val_del}{Delayed value to use for computing the indifference curve, if applicable}

\item{verbose}{Specifies whether to provide extra detail, if applicable}

\item{...}{Further arguments passed to `integrate()`}
}
\value{
AUC value
}
\description{
Compute the area under the curve using numerical integration
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bcnm(td_bc_single_ptpt)
print(AUC(mod))
}
}
