% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.td_bclm}
\alias{coef.td_bclm}
\title{Extract model coefficients}
\usage{
\method{coef}{td_bclm}(object, df_par = T, ...)
}
\arguments{
\item{object}{An object of class \code{td_bcnm}}

\item{df_par}{Boolean specifying whether the coefficients returned should be the parameters of a discount function (versus the beta parameters from the regression)}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of coefficients
}
\description{
Get coefficients of a temporal discounting binary choice model
}
