% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.td_ddm}
\alias{coef.td_ddm}
\title{Extract model coefficients}
\usage{
\method{coef}{td_ddm}(object, type = "all", ...)
}
\arguments{
\item{object}{An object of class \code{td_ddm}}

\item{type}{A string specifying which coefficients to extract. \code{'all'} extracts them all, \code{'ddm'} extracts only DDM-specific parameters, and \code{'df'} extracts only discount function parameters}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of coefficients
}
\description{
Get coefficients of a temporal discounting drift diffusion model
}
