% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.td_ipm}
\alias{coef.td_ipm}
\title{Extract model coefficients}
\usage{
\method{coef}{td_ipm}(mod)
}
\value{
A named vector of coefficients
}
\description{
Get coefficients of a temporal discounting indifference point model
}
