% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{deviance.td_bcnm}
\alias{deviance.td_bcnm}
\title{Model deviance}
\usage{
\method{deviance}{td_bcnm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{td_bcnm}.}

\item{...}{Additional arguments currently not used.}
}
\description{
Compute deviance for a temporal discounting binary choice model.
}
\seealso{
Other nonlinear binary choice model functions: 
\code{\link{coef.td_bcnm}()},
\code{\link{fitted.td_bcnm}()},
\code{\link{logLik.td_bcnm}()},
\code{\link{predict.td_bcnm}()},
\code{\link{residuals.td_bcnm}()},
\code{\link{td_bcnm}()}
}
\concept{nonlinear binary choice model functions}
