% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{deviance.td_ddm}
\alias{deviance.td_ddm}
\title{Model deviance}
\usage{
\method{deviance}{td_ddm}(mod)
}
\arguments{
\item{mod}{An object of class \code{td_ddm}}
}
\description{
Compute deviance for a temporal discounting drift diffusion model
}
