% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ED50}
\alias{ED50}
\title{Median effective delay}
\usage{
ED50(mod)
}
\arguments{
\item{mod}{A temporal discounting model.}
}
\value{
A vector of predictions
}
\description{
Compute the median effective delay
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bcm(td_bc_single_ptpt)
print(ED50(mod))
}
}
