% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ED50}
\alias{ED50}
\title{Median effective delay}
\usage{
ED50(mod)
}
\arguments{
\item{mod}{A temporal discounting model. See `td_gnlm`}

\item{newdata}{Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.}

\item{type}{The type of prediction required. As in predict.glm, `"link"` (default) and `"response"` give predictions on the scales of the linear predictors and response variable, respectively. `"indiff"` gives predicted indifference points. In this case, `newdata` needs only a `del` column.}
}
\value{
A vector of predictions
}
\description{
Compute the median effective delay
}
