% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.td_bcnm}
\alias{fitted.td_bcnm}
\title{Get fitted values}
\usage{
\method{fitted}{td_bcnm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{td_bcnm}}.}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of fitted values
}
\description{
Get fitted values of a temporal discounting binary choice model
}
\seealso{
Other nonlinear binary choice model functions: 
\code{\link{coef.td_bcnm}()},
\code{\link{deviance.td_bcnm}()},
\code{\link{logLik.td_bcnm}()},
\code{\link{predict.td_bcnm}()},
\code{\link{residuals.td_bcnm}()},
\code{\link{td_bcnm}()}
}
\concept{nonlinear binary choice model functions}
