% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.td_ddm}
\alias{fitted.td_ddm}
\title{Get fitted values}
\usage{
\method{fitted}{td_ddm}(object, ...)
}
\arguments{
\item{object}{An object of class \link{td_ddm}}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of fitted values
}
\description{
Get fitted values of a temporal discounting drift diffusion model
}
