% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.td_ipm}
\alias{fitted.td_ipm}
\title{Get fitted values}
\usage{
\method{fitted}{td_ipm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{td_ipm}}

\item{...}{Additional arguments currently not used.}
}
\value{
A named vector of fitted values
}
\description{
Get fitted values of a temporal discounting indifference point model
}
