% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indiffs}
\alias{indiffs}
\title{Get model-free indifference points}
\usage{
indiffs(mod)
}
\arguments{
\item{mod}{A model of class \code{\link{td_bcnm}}, \code{\link{td_ipm}}, or \code{\link{td_ddm}} for which the "model-free" discount function has been fit.}
}
\value{
A dataframe with columns \code{del} (delay) and \code{indiff} (indifference point).
}
\description{
Convert a temporal discounting model with the "model-free" discount function to a dataframe of indifference points.
}
\examples{
\donttest{
data("td_bc_single_ptpt")
mod <- td_bcnm(td_bc_single_ptpt, discount_function = 'model-free')
indiff_data <- indiffs(mod)
}
}
