% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kirby_score}
\alias{kirby_score}
\title{Kirby MCQ-style scoring}
\usage{
kirby_score(data, discount_function = c("hyperbolic", "exponential"))
}
\arguments{
\item{data}{Responses to score.}

\item{discount_function}{Should \eqn{k} values be computed according to the hyperbolic or exponential discount function? The original method uses the hyperbolic, but in principle the exponential is also possible.}
}
\value{
An object of class \code{\link{td_ipm}}.
}
\description{
Score a set of responses according to the method of Kirby et al. (1999, \doi{10.1037//0096-3445.128.1.78}). This is described in detail in Kaplan et al. (2016, \doi{10.1007/s40614-016-0070-9}).
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- kirby_score(td_bc_single_ptpt)
}
}
