% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.td_bcnm}
\alias{logLik.td_bcnm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{td_bcnm}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{td_bcnm}}

\item{...}{Additional arguments currently not used.}
}
\description{
Compute log-likelihood for a temporal discounting binary choice model.
}
\seealso{
Other nonlinear binary choice model functions: 
\code{\link{coef.td_bcnm}()},
\code{\link{deviance.td_bcnm}()},
\code{\link{fitted.td_bcnm}()},
\code{\link{predict.td_bcnm}()},
\code{\link{residuals.td_bcnm}()},
\code{\link{td_bcnm}()}
}
\concept{nonlinear binary choice model functions}
