% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.td_ddm}
\alias{logLik.td_ddm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{td_ddm}(object, type = c("resp_rt", "resp", "rt"), ...)
}
\arguments{
\item{type}{Should probabilities /probability densities be computed for responses and RTs (\code{'resp_rt'}, default) or responses only (\code{'resp'})?}

\item{...}{Additional arguments currently not used.}

\item{mod}{An object of class \code{td_bcnm}.}
}
\description{
Compute log-likelihood for a temporal discounting drift diffusion model.
}
\seealso{
Other drift diffusion model functions: 
\code{\link{coef.td_ddm}()},
\code{\link{deviance.td_ddm}()},
\code{\link{fitted.td_ddm}()},
\code{\link{predict.td_ddm}()},
\code{\link{td_ddm}()}
}
\concept{drift diffusion model functions}
