% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.td_um}
\alias{plot.td_um}
\title{Plot models}
\usage{
\method{plot}{td_um}(
  x,
  type = c("summary", "endpoints", "link"),
  legend = T,
  verbose = T,
  del = NULL,
  val_del = NULL,
  ...
)
}
\arguments{
\item{x}{A delay discounting model. See \code{dd_prob_model} and \code{dd_det_model}}

\item{type}{Type of plot to generate}

\item{legend}{Logical: display a legend? Ignored if \code{type != 'summary'}}

\item{verbose}{Whether to print info about, e.g., setting del = ED50 when type == 'endpoints'}

\item{del}{Plots data for a particular delay}

\item{val_del}{Plots data for a particular delayed value}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Plot delay discounting models
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
plot(mod, type = 'summary')
plot(mod, type = 'endpoints')
}
}
