% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.td_um}
\alias{plot.td_um}
\title{Plot models}
\usage{
\method{plot}{td_um}(mod, type = c("summary", "endpoints", "link"), ...)
}
\arguments{
\item{mod}{A delay discounting model. See `dd_prob_model` and `dd_det_model`}

\item{type}{Type of plot to generate}
}
\description{
Plot delay discounting models
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
plot(mod, type = 'summary')
plot(mod, type = 'endpoints')
}
}
