% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.td_um}
\alias{plot.td_um}
\title{Plot models}
\usage{
\method{plot}{td_um}(
  x,
  type = c("summary", "endpoints", "link", "rt"),
  legend = T,
  verbose = T,
  del = NULL,
  val_del = NULL,
  confint = 0.95,
  ...
)
}
\arguments{
\item{x}{A delay discounting model. See \code{dd_prob_model} and \code{dd_det_model}}

\item{type}{Type of plot to generate}

\item{legend}{Logical: display a legend? Only relevant for \code{type = 'summary'} and \code{type = 'rt'}}

\item{verbose}{Whether to print info about, e.g., setting del = ED50 when \code{type = 'endpoints'}}

\item{del}{Plots data for a particular delay}

\item{val_del}{Plots data for a particular delayed value}

\item{confint}{When \code{type = 'rt'}, what confidence interval should be plotted for RTs? Default is 0.95 (95\% confidence interval)}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Plot delay discounting models
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
plot(mod, type = 'summary')
plot(mod, type = 'endpoints')
}
}
