% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.td_um}
\alias{plot.td_um}
\title{Plot models}
\usage{
\method{plot}{td_um}(
  x,
  type = c("summary", "endpoints", "link", "rt"),
  legend = TRUE,
  p_lines = NULL,
  p_tol = 0.001,
  verbose = TRUE,
  del = NULL,
  val_del = NULL,
  q_lines = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{x}{A delay discounting model. See \code{dd_prob_model} and \code{dd_det_model}.}

\item{type}{Type of plot to generate.}

\item{legend}{Logical: display a legend? Only relevant for \code{type = 'summary'} and \code{type = 'rt'}.}

\item{p_lines}{Numerical vector. When \code{type = 'summary'} the discount curve, where the probability of selecting the immediate reward is 0.5, is plotted. \code{p_lines} allows you to specify other probabilities for which similar curves should be plotted (only applicable for probabilistic models, e.g. \code{td_bcnm}, \code{td_bclm} and \code{td_ddm}).}

\item{p_tol}{If \code{p_lines} is not \code{NULL}, what is the maximum distance that estimated probabilities can be from their true values? Smaller values results in slower plot generation.}

\item{verbose}{Whether to print info about, e.g., setting del = ED50 when \code{type = 'endpoints'}.}

\item{del}{Plots data for a particular delay.}

\item{val_del}{Plots data for a particular delayed value.}

\item{q_lines}{When \code{type = 'rt'}, plot dashed lines for which quantiles of the predicted RT distribution? Default is 0.025 and 0.975 (i.e., a quantile-based 95\% confidence interval).}

\item{...}{Additional arguments to \code{plot()}.}
}
\value{
No return value (called to produce a plot)
}
\description{
Plot delay discounting models.
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bclm(td_bc_single_ptpt, model = 'hyperbolic.1')
plot(mod, type = 'summary', p_lines = c(0.25, 0.75), log = 'x')
plot(mod, type = 'endpoints')
}
}
