% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.td_bcm}
\alias{predict.td_bcm}
\title{Model Predictions}
\usage{
\method{predict}{td_bcm}(mod, newdata = NULL, type = c("link", "response", "indiff"))
}
\arguments{
\item{mod}{A temporal discounting binary choice model. See `td_bcm`.}

\item{newdata}{Optionally, a data frame to use for prediction. If omitted, the data used to fit the model will be used for prediction.}

\item{type}{The type of prediction required. As in predict.glm, `"link"` (default) and `"response"` give predictions on the scales of the linear predictors and response variable, respectively. `"indiff"` gives predicted indifference points. In this case, `newdata` needs only a `del` column.}
}
\value{
A vector of predictions
}
\description{
Generate predictions from a temporal discounting binary choice model
}
\examples{
\dontrun{
data("td_bc_single_ptpt")
mod <- td_bcm(td_bc_single_ptpt, discount_function = 'hyperbolic')
indiffs <- predict(mod, newdata = data.frame(del = 1:100), type = 'indiff')
}
}
