% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.td_ipm}
\alias{predict.td_ipm}
\title{Model Predictions}
\usage{
\method{predict}{td_ipm}(mod, del = NULL, newdata = NULL, ...)
}
\arguments{
\item{mod}{A temporal discounting indifference point model. See `td_ipm`.}

\item{del}{Vector of delays for which to predict indifference points. If omitted, the data used to fit the model will be used for prediction.}

\item{newdata}{Optionally, a data frame to use for prediction. This overrides the `del` argument.}

\item{type}{The type of prediction required. As in predict.glm, `"link"` (default) and `"response"` give predictions on the scales of the linear predictors and response variable, respectively. `"indiff"` gives predicted indifference points. In this case, `newdata` needs only a `del` column.}
}
\value{
A vector of predictions
}
\description{
Generate predictions from a temporal discounting indifference point model
}
