% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.td_bcm}
\alias{residuals.td_bcm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{td_bcm}(object, ...)
}
\arguments{
\item{object}{A temporal discounting binary choice model. See \code{td_bcm}.}

\item{...}{Additional arguments currently not used.}

\item{type}{The type of residuals to be returned. See \code{residuals.glm}.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting binary choice model
}
