% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.td_ipm}
\alias{residuals.td_ipm}
\title{Residuals from temporal discounting model}
\usage{
\method{residuals}{td_ipm}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{A temporal discounting model. See \code{td_bcnm}.}

\item{type}{The type of residuals to be returned. See \code{residuals.nls}.}

\item{...}{Additional arguments currently not used.}
}
\value{
A vector of residuals
}
\description{
Get residuals from a temporal discounting indifference point model
}
\seealso{
Other indifference point model functions: 
\code{\link{coef.td_ipm}()},
\code{\link{fitted.td_ipm}()},
\code{\link{logLik.td_ipm}()},
\code{\link{predict.td_ipm}()},
\code{\link{td_ipm}()}
}
\concept{indifference point model functions}
