% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{td_bc_study}
\alias{td_bc_study}
\title{Binary choice data for a study}
\description{
Data from 421 participants, who each made 70 binary choices. Along with the columns required by \code{td_bcnm}, the reaction time (\code{rt}) is recorded. Participants are identified by the alphnumeric code in the \code{id} column.
}
\references{
\url{doi.org/10.31234/osf.io/y2fdh}
}
\author{
Isaac Kinley \email{isaac.kinley@gmail.com}
}
\keyword{choice}
\keyword{delay}
\keyword{discounting}
\keyword{temporal}
