% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_ipm.R
\name{td_ipm}
\alias{td_ipm}
\title{Indifference point model}
\usage{
td_ipm(
  data,
  discount_function = c("all", "hyperbolic", "exponential", "inverse-q-exponential",
    "nonlinear-time-hyperbolic", "scaled-exponential", "dual-systems-exponential",
    "nonlinear-time-exponential", "model-free", "noise"),
  optim_args = list(),
  silent = T
)
}
\arguments{
\item{data}{A data frame with columns `indiff` for the pre-computed indifference points and `del` for the delay}

\item{discount_function}{A vector of strings specifying the name of the discount functions to use, or an object of class `td_fn`.}

\item{silent}{A Boolean specifying whether the call to `optim` (which occurs in a `try` block) should be silent on error}
}
\value{
A list from `optim` with additional components specifying the AIC, the ED50, the discount function, and the probabilistic model
}
\description{
Compute a model of a single subject's indifference points
}
\note{
The `par` component of the output list is for internal use. For statistical analyses, use the `untransformed_parameters`. `par` contains the parameters after various transformations intended to keep them within certain bounds (e.g., k parameters should never be negative)
}
\examples{
# Generate data
df <- data.frame(del = exp(1:10), indiff = 1 / (1 + 0.001*exp(1:10)))
# Fit model
mod <- dd_det_model(df)
print(mod$discount_function)
}
