% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_ipm.R
\name{td_ipm}
\alias{td_ipm}
\title{Temporal discounting indifference point model}
\usage{
td_ipm(
  data,
  discount_function = c("all", "hyperbolic", "exponential", "inverse-q-exponential",
    "nonlinear-time-hyperbolic", "scaled-exponential", "dual-systems-exponential",
    "nonlinear-time-exponential", "model-free", "constant"),
  optim_args = list(),
  silent = TRUE
)
}
\arguments{
\item{data}{A data frame with columns \code{indiff} for the pre-computed indifference points and \code{del} for the delay.}

\item{discount_function}{A vector of strings specifying the name of the discount functions to use, or an object of class \code{td_fn} (used for creating custom discount functions), or a list of objects of class \code{td_fn}.}

\item{optim_args}{A list of additional args to pass to \code{optim}.}

\item{silent}{A Boolean specifying whether the call to \code{optim} (which occurs in a \code{try} block) should be silent on error.}
}
\value{
An object of class \code{td_ipm} with components \code{data} (containing the data used for fitting), \code{config} (containing the internal configuration of the model, including the \code{discount_function}), and \code{optim} (the output of \code{optim()}).
}
\description{
Compute a model of a single subject's indifference points.
}
\examples{
\dontrun{
# Basic usage
data("td_ip_simulated_ptpt")
mod <- td_ipm(td_ip_simulated_ptpt, discount_function = "hyperbolic")
# Custom discount function
custom_discount_function <- td_fn(
  name = 'custom',
  fn = function(data, p) (1 - p['b'])*exp(-p['k']*data$del) + p['b'],
  par_starts = list(k = c(0.001, 0.1), b = c(0.001, 0.1)),
  par_lims = list(k = c(0, Inf), b = c(0, 1)),
  ED50 = function(p) 'non-analytic'
)
mod <- td_ipm(td_ip_simulated_ptpt, discount_function = custom_discount_function)
}
}
\concept{indifference point model functions.}
