% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wileyto_score}
\alias{wileyto_score}
\title{Wileyto score a questionnaire}
\usage{
wileyto_score(data)
}
\arguments{
\item{data}{Responses to score.}
}
\value{
An object of class \code{td_bclm}.
}
\description{
Score a set of responses according to the method of Wileyto et al. (2004, \doi{10.3758/BF03195548}). This function is a thin wrapper to \code{\link{td_bclm}}.
}
\examples{
\donttest{
data("td_bc_single_ptpt")
mod <- wileyto_score(td_bc_single_ptpt)
}
}
